requirements.txt
absl-py==2.1.0
attrs==24.2.0
contourpy==1.2.1
cycler==0.12.1
flatbuffers==24.3.25
fonttools==4.53.1
importlib-resources==6.4.0
kiwisolver==1.4.5
load-dotenv==0.1.0
matplotlib==3.9.1.post1
mediapipe==0.9.1.0
numpy==2.0.1
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
packaging==24.1
pillow==10.4.0
protobuf==3.20.3
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
slack-sdk==3.31.0
zipp==3.20.0
